Notify parent focus delegates when a child changes focus.
FocusDelegate.resolve() can now get a SYNC_CHILD "direction", indicating that child focus may have changed and the delegate should
update state. Preferred resolve() to adding a new api. The parent updates will not dispatch blur or focus events.

Addresses a use case where a screen initially loads and sets focus on an item inside of a focus group. Focus is set, but the parent focus
group has a stale notion of focus. On the next user navigation, the focus would move to the wrong element.