Add rgba function support for color, backgroundColor, borderColor and tintColor.
rgba() works like the CSS version of the function. Specifying it on a Style will apply the alpha to the view only and prevent opacity from cascading
down the view graph.

As a shorthand, string hex values support an optional fourth component for alpha. #RRGGBBAA (or #RGBA) is equivalent to rgba(RR, GG, BB, AA).